+++ tests/testthat/test-conda.R
@@ -7,11 +7,42 @@ test_that("can install pre-commit", {
   expect_output(
     {
       git2r::init(path = tempdir)
-      use_precommit(tempdir, open = FALSE, force = TRUE)
+      use_precommit(open = FALSE, force = TRUE, path_root = tempdir)
     },
     "to get the latest"
   )
 })
+
+test_that("fails early if repo is not a git repo ", {
+  expect_error(
+    {
+      tempdir <- fs::path(tempdir(), "t9")
+      fs::dir_create(tempdir)
+      use_precommit(path_root = tempdir)
+    },
+    "is not a git repo"
+  )
+})
+
+test_that("can use custom config file ", {
+  tempdir <- fs::path(tempdir(), "t10")
+  fs::dir_create(tempdir)
+  tempdir2 <- fs::path(tempdir(), "t11")
+  fs::dir_create(tempdir2)
+  path_custom <- fs::path(tempdir2, "some-precommit.yaml")
+  new_text <- "# 4js93"
+  readLines(system.file("pre-commit-config.yaml", package = "precommit")) %>%
+    c(new_text) %>%
+    writeLines(path_custom)
+  git2r::init(tempdir)
+  use_precommit(path_cp_config_from = path_custom, path_root = tempdir, force = TRUE)
+  config <- readLines(fs::path(tempdir, ".pre-commit-config.yaml"))
+  expect_equal(
+    config[length(config)],
+    new_text
+  )
+})
+
 
 
 test_that("Can uninstall pre-commit (repo scope)", {
@@ -31,7 +62,7 @@ test_that("Can uninstall pre-commit (repo scope)", {
   )
 
   # when there is no pre-commit.yaml anymore
-  use_precommit(tempdir, open = FALSE, force = TRUE)
+  use_precommit(open = FALSE, force = TRUE, path_root = tempdir)
   fs::file_delete(fs::path(tempdir, ".pre-commit-config.yaml"))
   expect_output(
     uninstall_precommit(scope = "repo", path_root = tempdir),
@@ -48,15 +79,22 @@ test_that("Can uninstall (globally)", {
     uninstall_precommit(scope = "global", ask = "none"),
     "No installation found."
   )
-  expect_error(install_precommit(), NA)
 })
 
+test_that("use_precommit fails when no global installation is found", {
+  expect_error(use_precommit(path_root = tempdir), "installed on your system")
+})
 
-test_that("can use precommit", {
-  expect_error({
-    tempdir <- tempdir()
-    git2r::init(tempdir)
-    use_precommit(path_root = tempdir)
-  }, NA)
-  
+test_that("can install pre-commit with remote config", {
+  expect_error(install_precommit(), NA)
+
+  expect_output(
+    {
+      git2r::init(path = tempdir)
+      use_precommit(example_remote_config(),
+        open = FALSE, force = TRUE, path_root = tempdir
+      )
+    },
+    "to get the latest"
+  )
 })